+++ tests/e2e/landingPage.spec.ts
@@ -2,7 +2,8 @@ import { qase } from 'playwright-qase-reporter';
 
 import { EXCHANGE_TAB_LABEL_PATTERN, WALLET_OPTIONS } from './data';
 import { noWalletTest as test } from './fixtures/noWallet';
-import { ConnectWalletPage, LandingPage } from './pages';
+import { ConnectWalletPage } from './pages/ConnectWalletPage';
+import { LandingPage } from './pages/LandingPage';
 
 test.describe('Landing page and navigation', () => {
   test.beforeEach(async ({ page }) => {
