+++ tests/e2e/mainMenu.spec.ts
@@ -3,8 +3,9 @@ import { qase } from 'playwright-qase-reporter';
 
 import { URLS } from './data';
 import { noWalletTest as test } from './fixtures/noWallet';
-import { LandingPage, MainMenuPage, ScanPage } from './pages';
-
+import { LandingPage } from './pages/LandingPage';
+import { MainMenuPage } from './pages/MainMenuPage';
+import { ScanPage } from './pages/ScanPage';
 const NAV_TIMEOUT_MS = 30_000;
 
 test.describe('Main Menu flows', () => {
