+++ tests/e2e/mobileView.spec.ts
@@ -3,7 +3,8 @@ import { qase } from 'playwright-qase-reporter';
 
 import enTranslation from '../../src/i18n/translations/en/translation.json' with { type: 'json' };
 import { noWalletTest as test } from './fixtures/noWallet';
-import { LandingPage, MainMenuPage, Theme } from './pages';
+import { LandingPage } from './pages/LandingPage';
+import { MainMenuPage, Theme } from './pages/MainMenuPage';
 import { isFullyInViewport } from './utils/elementUtils';
 import { removeFormattingTags } from './utils/translationUtils';
