--- tests/e2e/pages/index.ts
@@ -1,9 +1,0 @@
-export { ConnectWalletPage } from './ConnectWalletPage';
-export { EarnPage } from './EarnPage';
-export { LandingPage } from './LandingPage';
-export { MainMenuPage, Theme } from './MainMenuPage';
-export { PortfolioPage } from './PortfolioPage';
-export { ProfilePage } from './ProfilePage';
-export { ScanPage } from './ScanPage';
-export { SettingsPage } from './SettingsPage';
-export { WidgetPage } from './WidgetPage';
